refactor(engine): Rename templates module to expressions
