+++ tracecat/expressions/__init__.py
@@ -1,5 +1,5 @@
+from .engine import TemplateExpression
 from .eval import eval_templated_object, extract_templated_secrets
-from .expressions import TemplateExpression
 from .validators import TemplateValidator
 
 __all__ = [
